requirements.txt
google-generativeai
python-frontmatter
python-dotenv
Pillow
